Move the Single Document Switch to the status bar
