+++ packages/statusbar-extension/src/index.ts
@@ -26,6 +26,8 @@ import { IDocumentWidget } from '@jupyterlab/docregistry';
 
 import { FileEditor, IEditorTracker } from '@jupyterlab/fileeditor';
 
+import { IMainMenu } from '@jupyterlab/mainmenu';
+
 import {
   INotebookTracker,
   Notebook,
@@ -41,11 +43,13 @@ import {
   StatusBar
 } from '@jupyterlab/statusbar';
 
-import { IMainMenu } from '@jupyterlab/mainmenu';
-
 import { ISettingRegistry } from '@jupyterlab/settingregistry';
 
 import { ITranslator } from '@jupyterlab/translation';
+
+import { Switch } from '@jupyterlab/ui-components';
+
+import { CommandRegistry } from '@lumino/commands';
 
 import { Title, Widget } from '@lumino/widgets';
 
@@ -381,12 +385,69 @@ export const runningSessionsItem: JupyterFrontEndPlugin<void> = {
   }
 };
 
+/**
+ * The single-document mode switch in the top area.
+ */
+const modeSwitch: JupyterFrontEndPlugin<void> = {
+  id: '@jupyterlab/application-extension:mode-switch',
+  requires: [ILabShell, ITranslator, IStatusBar],
+  activate: (
+    app: JupyterFrontEnd,
+    shell: ILabShell,
+    translator: ITranslator,
+    statusBar: IStatusBar
+  ) => {
+    const trans = translator.load('jupyterlab');
+    const modeSwitch = new Switch();
+    modeSwitch.id = 'jp-single-document-mode';
+
+    modeSwitch.valueChanged.connect((_, args) => {
+      shell.mode = args.newValue ? 'single-document' : 'multiple-document';
+    });
+    shell.modeChanged.connect((_, mode) => {
+      modeSwitch.value = mode === 'single-document';
+    });
+    modeSwitch.value = shell.mode === 'single-document';
+
+    // Show the current file browser shortcut in its title.
+    const updateModeSwitchTitle = () => {
+      const binding = app.commands.keyBindings.find(
+        b => b.command === 'application:toggle-mode'
+      );
+      if (binding) {
+        const ks = CommandRegistry.formatKeystroke(binding.keys.join(' '));
+        modeSwitch.caption = trans.__('Single-Document Mode (%1)', ks);
+      } else {
+        modeSwitch.caption = trans.__('Single-Document Mode');
+      }
+    };
+    updateModeSwitchTitle();
+    app.commands.keyBindingChanged.connect(() => {
+      updateModeSwitchTitle();
+    });
+
+    modeSwitch.label = trans.__('Mode');
+
+    statusBar.registerStatusItem(
+      '@jupyterlab/application-extension:mode-switch',
+      {
+        item: modeSwitch,
+        align: 'left',
+        isActive: () => true,
+        rank: -1
+      }
+    );
+  },
+  autoStart: true
+};
+
 const plugins: JupyterFrontEndPlugin<any>[] = [
   statusBar,
   lineColItem,
   kernelStatus,
   runningSessionsItem,
-  memoryUsageItem
+  memoryUsageItem,
+  modeSwitch
 ];
 
 export default plugins;
