+++ packages/statusbar-extension/tsconfig.json
@@ -38,6 +38,9 @@
     },
     {
       "path": "../translation"
+    },
+    {
+      "path": "../ui-components"
     }
   ]
 }
